+++ backend/src/main/java/com/woowacourse/gongseek/article/application/dto/ArticleIdResponse.java
@@ -1,4 +1,4 @@
-package com.woowacourse.gongseek.article.presentation.dto;
+package com.woowacourse.gongseek.article.application.dto;
 
 import com.woowacourse.gongseek.article.domain.Article;
 import lombok.AccessLevel;
@@ -9,13 +9,11 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
-public class ArticleUpdateResponse {
+public class ArticleIdResponse {
 
     private Long id;
-    private String category;
 
-    public ArticleUpdateResponse(Article article) {
+    public ArticleIdResponse(Article article) {
         this.id = article.getId();
-        this.category = article.getCategory().getValue();
     }
 }
