+++ backend/src/main/java/com/woowacourse/gongseek/article/application/dto/ArticleUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.gongseek.article.presentation.dto;
+package com.woowacourse.gongseek.article.application.dto;
 
 import java.util.List;
 import lombok.AccessLevel;
